Extract Options into its own file.
Register Handlers specifically for the services used based on the Options vs registering all using the AddMediatR extention.
Cleaned up namespaces for ReduxDevTools.  At some point a folder doesn't require a namespace.
Made all the Payloads nested classes as they make no sense outside that class anyway.
