+++ source/BlazorState/Pipeline/ReduxDevTools/Features/IReduxRequest.cs
@@ -1,8 +1,7 @@
-﻿namespace BlazorState.Pipeline.ReduxDevTools.Features
+﻿namespace BlazorState.Pipeline.ReduxDevTools
 {
   /// <summary>
   /// Marker Interface to allow for filtering of Devtools Requests
   /// </summary>
-  internal interface IReduxRequest
-  { }
+  internal interface IReduxRequest { }
 }
