+++ source/BlazorState/Pipeline/ReduxDevTools/Features/JumpToState/JumpToStateHandler.cs
@@ -1,4 +1,4 @@
-﻿namespace BlazorState.Pipeline.ReduxDevTools.Features.JumpToState
+﻿namespace BlazorState.Pipeline.ReduxDevTools
 {
   using BlazorState;
   using MediatR;
