+++ source/BlazorState/Pipeline/ReduxDevTools/Features/JumpToState/JumpToStateRequest.cs
@@ -1,13 +1,15 @@
-﻿namespace BlazorState.Pipeline.ReduxDevTools.Features.JumpToState
+﻿namespace BlazorState.Pipeline.ReduxDevTools
 {
   using MediatR;
 
-  internal class JumpToStateRequest : DispatchRequest<Payload>, IRequest, IReduxRequest { }
-
-  internal class Payload
+  internal class JumpToStateRequest : DispatchRequest<JumpToStateRequest.Payload>, IRequest, IReduxRequest
   {
-    public int ActionId { get; set; }
-    public string Type { get; set; }
 
+    internal class Payload
+    {
+      public int ActionId { get; set; }
+      public string Type { get; set; }
+
+    }
   }
 }
