+++ source/BlazorState/Pipeline/ReduxDevTools/Features/Start/StartRequest.cs
@@ -1,11 +1,12 @@
-﻿namespace BlazorState.Pipeline.ReduxDevTools.Features.Start
+﻿namespace BlazorState.Pipeline.ReduxDevTools
 {
   using MediatR;
 
   /// <summary>
   /// Request received from Redux Dev Tools when one presses the Start Button.
   /// </summary>
-  internal class StartRequest : DispatchRequest<Payload>, IRequest, IReduxRequest { }
+  internal class StartRequest : DispatchRequest<StartRequest.Payload>, IRequest, IReduxRequest {
 
   internal class Payload { }
+  }
 }
